third_party/blink/web_tests/virtual/focusless-spat-nav/fast/spatial-navigation/focusless/snav-focusless-fully-contained-input.html

<!DOCTYPE html>
<script src="../../../../../resources/testharness.js"></script>
<script src="../../../../../resources/testharnessreport.js"></script>
<script src="file:///gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="file:///gen/third_party/blink/public/mojom/page/spatial_navigation.mojom.js"></script>
<script src="../../../../../fast/spatial-navigation/resources/mock-snav-service.js"></script>
<script src="../../../../../fast/spatial-navigation/resources/snav-testharness.js"></script>

<style>
  div {
    width: 200px;
    height: 200px;
    margin: 5px;
    border: 1px solid black;
  }

  input {
    margin: 5px;
  }
</style>

<div tabindex="0">
  <input id="input" type="text" value="second"></input>
</div>

<script>
  const input = document.getElementById("input");

  // Spatial navigation attempts to determine when one valid target fully
  // contains another. In this case, the contained "inner" target should be
  // selected. This is done by performing a hit test.
  //
  // This test ensures the hit test works correctly on an input box. This
  // requires special attention because the hit element will be in the input's
  // shadow dom.
  test(() => {
    assert_true(!!window.internals);
    snav.triggerMove('Down');
    assert_equals(window.internals.interestedElement,
                  input,
                  "Expected interest to move to input box.");
  }, "Navigation to fully contained input box.");
</script>
